svn 5968

+++ sapfor/experts/Sapfor_2017/_src/DynamicAnalysis/gCov_parser_func.h
@@ -9,5 +9,5 @@
 
 void parse_gcovfile(SgFile *file, const std::string &basefileName, std::map<int, Gcov_info> &gCovInfo, bool keep);
 bool __gcov_doesThisLineExecuted(const std::string &file, const int line);
-std::pair<int, int> __gcov_GetExecuted(const std::string &file, const int line);
+std::pair<int, int64_t> __gcov_GetExecuted(const std::string &file, const int line);
 void parseTimesDvmStatisticFile(const std::string &file, std::map<std::string, std::map<int, double>> &timesFromDvmStat);
